FEM-realtime: raw websocket implementation
