+++ 03-frontend-masters/11-intro-to-realtime/websockets/exercise-raw/backend/server.js
@@ -24,11 +24,55 @@ const server = http.createServer((request, response) => {
   });
 });
 
-/*
- *
- * your code goes here
- *
- */
+// websocket connection
+server.on("upgrade", (req, socket) => {
+  if (req.headers["upgrade"] !== "websocket") {
+    socket.end("HTTP/1.1 400 Bad Request");
+    return;
+  }
+
+  const acceptKey = req.headers["sec-websocket-key"];
+  const acceptValue = generateAcceptValue(acceptKey);
+
+  const headers = [
+    "HTTP/1.1 101 Web Socket Protocol Handshake",
+    "Upgrade: WebSocket",
+    "Connection: Upgrade",
+    `Sec-WebSocket-Accept: ${acceptValue}`,
+    "Sec-WebSocket-Protocol: json",
+    "\r\n", //END OF HEADERS
+  ];
+
+  socket.write(headers.join("\r\n"));
+  socket.write(objToResponse({ msg: getMsgs() }));
+
+  connections.push(socket);
+
+  // receiving data
+  socket.on("data", (buffer) => {
+    const message = parseMessage(buffer);
+
+    if (message) {
+      msg.push({
+        user: message.user,
+        text: message.text,
+        time: Date.now(),
+      });
+
+      connections.forEach((s) => {
+        s.write(objToResponse({ msg: getMsgs() }));
+      });
+    } else if (message === null) {
+      socket.end();
+    }
+  });
+
+  socket.on("end", () => {
+    connections = connections.filter((s) => s !== socket);
+  });
+});
+
+//
 
 const port = process.env.PORT || 8080;
 server.listen(port, () =>
