+++ 03-frontend-masters/11-intro-to-realtime/websockets/exercise-raw/frontend/raw-chat.js
@@ -11,14 +11,29 @@ chat.addEventListener("submit", function (e) {
 });
 
 async function postNewMsg(user, text) {
-  // code goes here
+  const data = { user, text };
+
+  ws.send(JSON.stringify(data));
 }
 
-/*
- *
- * your code goes here
- *
- */
+//
+const ws = new WebSocket("ws://localhost:8080", ["json"]);
+
+ws.addEventListener("open", () => {
+  console.log("Connected");
+  presence.innerText = "🎾";
+});
+
+ws.addEventListener("message", (e) => {
+  const data = JSON.parse(e.data);
+  allChat = data.msg;
+  render();
+});
+
+ws.addEventListener("close", () => {
+  presence.innerText = "🏓";
+});
+//
 
 function render() {
   const html = allChat.map(({ user, text }) => template(user, text));
