--- 03-frontend-masters/11-intro-to-realtime/websockets/raw/backend/generate-accept-value.js
@@ -1,13 +1,0 @@
-import crypto from "crypto";
-
-function generateAcceptValue(acceptKey) {
-  return (
-    crypto
-      .createHash("sha1")
-      // this magic string key is actually in the spec
-      .update(acceptKey + "258EAFA5-E914-47DA-95CA-C5AB0DC85B11", "binary")
-      .digest("base64")
-  );
-}
-
-export default generateAcceptValue;
