--- 03-frontend-masters/11-intro-to-realtime/websockets/raw/frontend/raw-chat.js
@@ -1,44 +1,0 @@
-const chat = document.getElementById("chat");
-const msgs = document.getElementById("msgs");
-const presence = document.getElementById("presence-indicator");
-let allChat = [];
-
-const ws = new WebSocket("ws://localhost:8080", ["json"]);
-
-ws.addEventListener("open", () => {
-  console.log("connected");
-  presence.innerText = "🟢";
-});
-
-ws.addEventListener("close", () => {
-  presence.innerText = "🔴";
-});
-
-ws.addEventListener("message", (event) => {
-  const data = JSON.parse(event.data);
-  allChat = data.msg;
-  render();
-});
-
-chat.addEventListener("submit", function (e) {
-  e.preventDefault();
-  postNewMsg(chat.elements.user.value, chat.elements.text.value);
-  chat.elements.text.value = "";
-});
-
-async function postNewMsg(user, text) {
-  const data = {
-    user,
-    text,
-  };
-
-  ws.send(JSON.stringify(data));
-}
-
-function render() {
-  const html = allChat.map(({ user, text }) => template(user, text));
-  msgs.innerHTML = html.join("\n");
-}
-
-const template = (user, msg) =>
-  `<li class="collection-item"><span class="badge">${user}</span>${msg}</li>`;
