--- 03-frontend-masters/11-intro-to-realtime/websockets/raw/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "raw-websocket",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node backend/server.js",
-    "dev": "nodemon backend/server.js"
-  },
-  "keywords": [],
-  "type": "module",
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "body-parser": "^1.19.0",
-    "express": "^4.17.1",
-    "morgan": "^1.10.0",
-    "nanobuffer": "^2.0.0",
-    "serve-handler": "^6.1.3"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.12"
-  }
-}
